% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmt_functions.R
\name{get_pmt_growth}
\alias{get_pmt_growth}
\title{Calculate first payment of a growth annuity due}
\usage{
get_pmt_growth(rate, growth, t)
}
\arguments{
\item{rate}{Annual discount rate (scalar double).}

\item{growth}{Annual growth rate (scalar double).}

\item{t}{Number of years in the future (scalar).}
}
\value{
numeric value.
}
\description{
With interest rate (rate), growth rate (g), and remaining period (t).
Payments are made in advance (beginning of each time period).
}
\examples{
get_pmt_growth(0.05, 0.02, 5) #0.2117597
}
