% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_pv}
\alias{get_pv}
\title{get_pv calculate the discount factor with rate and t}
\usage{
get_pv(rate, t)
}
\arguments{
\item{rate}{Annual discount rate (scalar double).}

\item{t}{Number of years in the future (scalar).}
}
\value{
Numeric value.
}
\description{
get_pv calculate the discount factor with rate and t
}
\examples{
get_pv(0.05,5) # 0.7835
}
