% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_pv_cf}
\alias{get_pv_cf}
\title{get_pv_cf calculates the present value of future cashflows (cf)}
\usage{
get_pv_cf(rate, cf)
}
\arguments{
\item{rate}{Annual discount rate (scalar double).}

\item{cf}{A list of cash flows in n=length(cf) number of years}
}
\value{
Numeric value.
}
\description{
get_pv_cf calculates the present value of future cashflows (cf)
}
\examples{
cf <- c(100, 200, 300, 400, 500, 600)
get_pv_cf(0.05, cf)  # 1704.37
}
