% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_pv_cf_roll}
\alias{get_pv_cf_roll}
\title{get_pv_cf_roll returns the remaining present value of future cash flows (cf) in every year forward}
\usage{
get_pv_cf_roll(rate, cf)
}
\arguments{
\item{rate}{Annual discount rate (scalar double).}

\item{cf}{A list of cash flows in n=length(cf) number of years}
}
\value{
A list of numeric values
}
\description{
get_pv_cf_roll returns the remaining present value of future cash flows (cf) in every year forward
}
\examples{
cf <- c(100, 200, 300, 400, 500, 600)
get_pv_cf_roll (0.05, cf) #$1,704.37 $1,609.13 $1,427.72 $1,168.57 $839.49) $447.73
}
