% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_pv_gpmt}
\alias{get_pv_gpmt}
\title{get_pv_gpmt calculates the present value of a growth annuity for 1 unit (such as $1) with interest rate, growth rate and t
Reference: https://financeformulas.net/Present_Value_of_Growing_Annuity.html}
\usage{
get_pv_gpmt(rate, growth, t)
}
\arguments{
\item{rate}{Annual discount rate (scalar double).}

\item{growth}{Annual growth rate (scalar double).}

\item{t}{Number of years in the future (scalar).}
}
\value{
Numeric value.
}
\description{
get_pv_gpmt calculates the present value of a growth annuity for 1 unit (such as $1) with interest rate, growth rate and t
Reference: https://financeformulas.net/Present_Value_of_Growing_Annuity.html
}
\examples{
get_pv_gpmt(0.08, 0.05, 18) # 13.2582
}
