% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_pv_pmt}
\alias{get_pv_pmt}
\title{get_pv_pmt calculates the present value for an annuity (same payment at equal intervals of time) for 1 unit (such as $1) with rate and t
Reference:financial mathematics for actuaries 3rd (p.21)
Link: https://actuarialscience.yolasite.com/resources/Ruckman.pdf}
\usage{
get_pv_pmt(rate, t)
}
\arguments{
\item{rate}{Annual discount rate (scalar double).}

\item{t}{Number of years in the future (scalar).}
}
\value{
Numeric value.
}
\description{
get_pv_pmt calculates the present value for an annuity (same payment at equal intervals of time) for 1 unit (such as $1) with rate and t
Reference:financial mathematics for actuaries 3rd (p.21)
Link: https://actuarialscience.yolasite.com/resources/Ruckman.pdf
}
\examples{
get_pv_pmt(.05, 5) # 4.3295
}
