% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pv_functions.R
\name{npv}
\alias{npv}
\title{#' Calculate the Present Value of Future Benefits (PVFB)
#'
#' Given a vector of separation rates, a vector of corresponding interest rates, and a vector of
#' future values (benefits), this function computes the present value of these benefits. At each
#' period \code{i}, it calculates the probability of separation and then discounts the subsequent
#' future values back to the present using the provided interest rates.
#'
#' @param sep_rate_vec Numeric vector. The annual separation rates for each future period.
#' @param interest_vec Numeric vector. The annual interest (discount) rates use in that period.
#' @param value_vec Numeric vector. The future benefits payable at each period.
#'
#' @return A numeric vector of the same length as \code{value_vec}, where each element represents the
#'         present value of future benefits starting from that period.
#' @export
#'
#' @examples
#' sep_rate_vec <- c(0.01, 0.02, 0.03, 0.04)
#' interest_vec <- c(0.05, 0.05, 0.05, 0.05)
#' value_vec <- c(100, 200, 300, 400)
#' get_pvfb(sep_rate_vec, interest_vec, value_vec)
get_pvfb <- function(sep_rate_vec, interest_vec, value_vec) {
N <- length(value_vec)
PVFB <- double(length = N)
for (i in 1:N) {
sep_rate <- sep_rate_vec\link{i:N}
sep_prob <- cumprod(1 - sep_rate) * sep_rate       # Probability of separating in each subsequent period
interest <- interest_vec\link{i}
if (i < N) {
value_sub <- value_vec\link{(i+1):N}                  # Payment in t+1 until the end of periods
sep_prob_sub <- sep_prob\link{-1}                     # Probability of remaining in the plan until the period t
df_sub <- (1 + interest)^(-(1:length(value_sub))) # Discount factors in each year based on the interest rate used in t
PVFB\link{i} <- sum(value_sub * sep_prob_sub * df_sub) # The product of probability, discount factor, future values (benefits) is PVFB
} else {
PVFB\link{i = N} <- 0                                 # At the last period, there are no future periods, so PVFB is 0
}
}
return(PVFB)
}
#' Calculate Present Value of Future Salaries (PVFS)
#'
#' This function computes the present value of future salaries (PVFS) for a given set of probabilities, interest rates, and salary values over a time period.
#'
#' @param remaining_prob_vec Numeric vector. The remaining survival probabilities for each period.
#' @param interest_vec Numeric vector. The interest rates for each period.
#' @param sal_vec Numeric vector. The projected salaries for each period.
#'
#' @return Numeric vector. A vector of PVFS values, where each element represents the present value of future salaries starting from that period.
#'
#' @details
#' The present value of future salaries is calculated by:
#' - Adjusting projected salaries by the survival probability at each period.
#' - Discounting future salaries to their present value using the interest rates provided.
#'
#' @examples
#' # Example inputs
#' remaining_prob_vec <- c(1, 0.95, 0.90, 0.85)
#' interest_vec <- c(0.03, 0.03, 0.03, 0.03)
#' sal_vec <- c(50000, 52000, 54000, 56000)
#'
#' # Calculate PVFS
#' get_pvfs(remaining_prob_vec, interest_vec, sal_vec)
#'
#' @export
get_pvfs <- function(remaining_prob_vec,
interest_vec,
sal_vec) {
N <- length(sal_vec)
PVFS <- double(length = N)
for (i in 1:N) {
remaining_prob_sub <- remaining_prob_vec\link{i:N} / remaining_prob_vec\link{i} # Calculate survival probabilities for future periods, using i year survival rate as the base
interest <- interest_vec\link{i}                                           # Get the interest rate for the current period
sal_sub <- sal_vec\link{i:N}                                               # Subset salaries for future periods from i period
df_sub  <- (1 + interest)^(-(1:length(sal_sub)))                      # Discount factors in each year based on the interest rate used in t
PVFS\link{i} <- sum(sal_sub * remaining_prob_sub * df_sub)                 # The sum of product of the future salaries, survival probability, and discount factor is present value of future salaries
}
return(PVFS)
}
npv calculates the present value of future cashflows (cf)}
\usage{
npv(rate, cf)
}
\arguments{
\item{rate}{Annual discount rate (scalar double).}

\item{cf}{A list of cash flows in n=length(cf) number of years}
}
\value{
Numeric value.
}
\description{
#' Calculate the Present Value of Future Benefits (PVFB)
#'
#' Given a vector of separation rates, a vector of corresponding interest rates, and a vector of
#' future values (benefits), this function computes the present value of these benefits. At each
#' period \code{i}, it calculates the probability of separation and then discounts the subsequent
#' future values back to the present using the provided interest rates.
#'
#' @param sep_rate_vec Numeric vector. The annual separation rates for each future period.
#' @param interest_vec Numeric vector. The annual interest (discount) rates use in that period.
#' @param value_vec Numeric vector. The future benefits payable at each period.
#'
#' @return A numeric vector of the same length as \code{value_vec}, where each element represents the
#'         present value of future benefits starting from that period.
#' @export
#'
#' @examples
#' sep_rate_vec <- c(0.01, 0.02, 0.03, 0.04)
#' interest_vec <- c(0.05, 0.05, 0.05, 0.05)
#' value_vec <- c(100, 200, 300, 400)
#' get_pvfb(sep_rate_vec, interest_vec, value_vec)
get_pvfb <- function(sep_rate_vec, interest_vec, value_vec) {
N <- length(value_vec)
PVFB <- double(length = N)
for (i in 1:N) {
sep_rate <- sep_rate_vec\link{i:N}
sep_prob <- cumprod(1 - sep_rate) * sep_rate       # Probability of separating in each subsequent period
interest <- interest_vec\link{i}
if (i < N) {
value_sub <- value_vec\link{(i+1):N}                  # Payment in t+1 until the end of periods
sep_prob_sub <- sep_prob\link{-1}                     # Probability of remaining in the plan until the period t
df_sub <- (1 + interest)^(-(1:length(value_sub))) # Discount factors in each year based on the interest rate used in t
PVFB\link{i} <- sum(value_sub * sep_prob_sub * df_sub) # The product of probability, discount factor, future values (benefits) is PVFB
} else {
PVFB\link{i = N} <- 0                                 # At the last period, there are no future periods, so PVFB is 0
}
}
return(PVFB)
}
#' Calculate Present Value of Future Salaries (PVFS)
#'
#' This function computes the present value of future salaries (PVFS) for a given set of probabilities, interest rates, and salary values over a time period.
#'
#' @param remaining_prob_vec Numeric vector. The remaining survival probabilities for each period.
#' @param interest_vec Numeric vector. The interest rates for each period.
#' @param sal_vec Numeric vector. The projected salaries for each period.
#'
#' @return Numeric vector. A vector of PVFS values, where each element represents the present value of future salaries starting from that period.
#'
#' @details
#' The present value of future salaries is calculated by:
#' - Adjusting projected salaries by the survival probability at each period.
#' - Discounting future salaries to their present value using the interest rates provided.
#'
#' @examples
#' # Example inputs
#' remaining_prob_vec <- c(1, 0.95, 0.90, 0.85)
#' interest_vec <- c(0.03, 0.03, 0.03, 0.03)
#' sal_vec <- c(50000, 52000, 54000, 56000)
#'
#' # Calculate PVFS
#' get_pvfs(remaining_prob_vec, interest_vec, sal_vec)
#'
#' @export
get_pvfs <- function(remaining_prob_vec,
interest_vec,
sal_vec) {
N <- length(sal_vec)
PVFS <- double(length = N)
for (i in 1:N) {
remaining_prob_sub <- remaining_prob_vec\link{i:N} / remaining_prob_vec\link{i} # Calculate survival probabilities for future periods, using i year survival rate as the base
interest <- interest_vec\link{i}                                           # Get the interest rate for the current period
sal_sub <- sal_vec\link{i:N}                                               # Subset salaries for future periods from i period
df_sub  <- (1 + interest)^(-(1:length(sal_sub)))                      # Discount factors in each year based on the interest rate used in t
PVFS\link{i} <- sum(sal_sub * remaining_prob_sub * df_sub)                 # The sum of product of the future salaries, survival probability, and discount factor is present value of future salaries
}
return(PVFS)
}
npv calculates the present value of future cashflows (cf)
}
\examples{
cf <- c(100, 200, 300, 400, 500, 600)
npv(0.05, cf)  # 1704.37
}
