% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/reason_functions_not_yet_implemented_with_new_code.R
\name{recur_grow}
\alias{recur_grow}
\title{#' @export
get_pvfs <- function(remaining_prob_vec, interest_vec, sal_vec) {
PVFS <- double(length = length(sal_vec))
for (i in 1:length(sal_vec)) {
remaining_prob_og <- remaining_prob_vec\link{i:length(remaining_prob_vec)}
remaining_prob <- remaining_prob_og / remaining_prob_og\link{1}
interest <- interest_vec\link{i}
sal <- sal_vec\link{i:length(sal_vec)}
sal_adjusted <- sal * remaining_prob
PVFS\link{i} <- npv(interest, sal_adjusted)
}
return(PVFS)
}}
\usage{
recur_grow(x, g)
}
\description{
#' @export
get_pvfs <- function(remaining_prob_vec, interest_vec, sal_vec) {
PVFS <- double(length = length(sal_vec))
for (i in 1:length(sal_vec)) {
remaining_prob_og <- remaining_prob_vec\link{i:length(remaining_prob_vec)}
remaining_prob <- remaining_prob_og / remaining_prob_og\link{1}
interest <- interest_vec\link{i}
sal <- sal_vec\link{i:length(sal_vec)}
sal_adjusted <- sal * remaining_prob
PVFS\link{i} <- npv(interest, sal_adjusted)
}
return(PVFS)
}
}
