% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_from_frs_model.R
\name{annfactor}
\alias{annfactor}
\title{#Present Value of Future Salaries (PVFS) function (to be applied to a vector of salaries)
#remaining_prob_vec is a vector containing the remaining probabilities. interest_vec is a discount rate (ARR) vector. sal_vec is a vector containing the salaries.
#' @export
get_pvfs <- function(remaining_prob_vec, interest_vec, sal_vec) {
PVFS <- double(length = length(sal_vec))
for (i in 1:length(sal_vec)) {
remaining_prob_og <- remaining_prob_vec\link{i:length(remaining_prob_vec)}
remaining_prob <- remaining_prob_og / remaining_prob_og\link{1}
interest <- interest_vec\link{i}
sal <- sal_vec\link{i:length(sal_vec)}
sal_adjusted <- sal * remaining_prob
PVFS\link{i} <- npv(interest, sal_adjusted)
}
return(PVFS)
}}
\usage{
annfactor(surv_DR_vec, cola_vec, one_time_cola = F)
}
\description{
#Present Value of Future Salaries (PVFS) function (to be applied to a vector of salaries)
#remaining_prob_vec is a vector containing the remaining probabilities. interest_vec is a discount rate (ARR) vector. sal_vec is a vector containing the salaries.
#' @export
get_pvfs <- function(remaining_prob_vec, interest_vec, sal_vec) {
PVFS <- double(length = length(sal_vec))
for (i in 1:length(sal_vec)) {
remaining_prob_og <- remaining_prob_vec\link{i:length(remaining_prob_vec)}
remaining_prob <- remaining_prob_og / remaining_prob_og\link{1}
interest <- interest_vec\link{i}
sal <- sal_vec\link{i:length(sal_vec)}
sal_adjusted <- sal * remaining_prob
PVFS\link{i} <- npv(interest, sal_adjusted)
}
return(PVFS)
}
}
