% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{annfactor}
\alias{annfactor}
\title{Calculate Annuity Factors with Survival and Cost-of-Living Adjustments}
\usage{
annfactor(surv_DR_vec, cola_vec, one_time_cola = F)
}
\arguments{
\item{surv_DR_vec}{Numeric vector. A vector of survival discount rates, representing the probabilities of survival for each period.}

\item{cola_vec}{Numeric vector. A vector of cost-of-living adjustment (COLA) rates for each period.}

\item{one_time_cola}{Logical, optional. If TRUE, a one-time COLA is applied (default is FALSE), this is not coded as no COLAs.}
}
\value{
Numeric vector. A vector of annuity factors, considering COLAs where each element corresponds to a period's annuity factor.
}
\description{
This function computes annuity factors for a series of survival discount rates, incorporating cost-of-living adjustments (COLA). It optionally supports a one-time COLA adjustment.
}
\examples{
# Example with survival discount rates and COLA rates
surv_DR_vec <- c(1.0, 0.9, 0.8, 0.7)
cola_vec <- c(0.02, 0.02, 0.02, 0.02)
annfactor(surv_DR_vec, cola_vec)

# Example with a one-time COLA
annfactor(surv_DR_vec, cola_vec, one_time_cola = FALSE)

}
