% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reason_functions_not_yet_implemented.R
\name{get_pmt0}
\alias{get_pmt0}
\title{#' npv calculates the present value of future cashflows (cf)
#'
#' @param rate Annual discount rate (scalar double).
#' @param cf A list of cash flows in n=length(cf) number of years
#'
#' @return Numeric value.
#' @export
#'
#' @examples
#' cf <- c(100, 200, 300, 400, 500, 600)
#' npv(0.05, cf)  # 1704.37
#' @export
npv <- function(rate, cf) {
df <- (1+rate)^(-(1:(length(cf))))    # Discount factor in each year based on rate
pv <- sum(cf * df)                    # The sum of the product of cash flow and discount factor in each year is PV
return(pv)
}
#Annuity factor for current retirees' benefits
#We need this function to calculate the annuity factors when a constant COLA is granted after the first year
#' @export}
\usage{
get_pmt0(r, nper, pv)
}
\description{
#' npv calculates the present value of future cashflows (cf)
#'
#' @param rate Annual discount rate (scalar double).
#' @param cf A list of cash flows in n=length(cf) number of years
#'
#' @return Numeric value.
#' @export
#'
#' @examples
#' cf <- c(100, 200, 300, 400, 500, 600)
#' npv(0.05, cf)  # 1704.37
#' @export
npv <- function(rate, cf) {
df <- (1+rate)^(-(1:(length(cf))))    # Discount factor in each year based on rate
pv <- sum(cf * df)                    # The sum of the product of cash flow and discount factor in each year is PV
return(pv)
}
#Annuity factor for current retirees' benefits
#We need this function to calculate the annuity factors when a constant COLA is granted after the first year
#' @export
}
