% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_pmt0}
\alias{get_pmt0}
\title{get_pmt0 calculates the total first payment of an annuity due with a present value (pv),
interest rate (r), and the number of periods (nper).
This function multiplies the payment factor (calculated using get_pmt_due) by the present value.}
\usage{
get_pmt0(r, nper, pv)
}
\arguments{
\item{r}{Annual discount rate (scalar double).}

\item{nper}{Number of periods (scalar integer).}

\item{pv}{Present value of the annuity (scalar numeric).}
}
\value{
numeric value representing the total first payment.
}
\description{
get_pmt0 calculates the total first payment of an annuity due with a present value (pv),
interest rate (r), and the number of periods (nper).
This function multiplies the payment factor (calculated using get_pmt_due) by the present value.
}
\examples{
get_pmt0(0.05, 5, 1000) # Example: calculates the total first payment
}
