% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_pmt_due}
\alias{get_pmt_due}
\title{get_pmt_due calculates the first payment of an annuity due with a present value pv, interest rate (rate), and remaining period (t)
payments are made in advance (beginning of each time period)
Reference: Annuity Due Payment - PV, https://financeformulas.net/Annuity-Due-Payment-from-Present-Value.html
Title}
\usage{
get_pmt_due(rate, t)
}
\arguments{
\item{rate}{Annual discount rate (scalar double).}

\item{t}{Number of years in the future (scalar).}
}
\value{
numeric value.
}
\description{
get_pmt_due calculates the first payment of an annuity due with a present value pv, interest rate (rate), and remaining period (t)
payments are made in advance (beginning of each time period)
Reference: Annuity Due Payment - PV, https://financeformulas.net/Annuity-Due-Payment-from-Present-Value.html
Title
}
\examples{
get_pmt_due(0.05, 5) # 0.219976
}
