% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_pmt_growth}
\alias{get_pmt_growth}
\title{First Payment for a Growing Annuity Due (Unit PV)}
\usage{
get_pmt_growth(rate, growth, t)
}
\arguments{
\item{rate}{Annual discount rate (scalar double).}

\item{growth}{Annual growth rate (scalar double).}

\item{t}{Number of years in the future (scalar).}
}
\value{
numeric value.
}
\description{
get_pmt_growth calculates the first payment for an growth annuity due with a present value pv,
interet rate (rate), # remaining period (t), and growth rate (g)
payments are made in advance (beginning of each time period)
}
\examples{
get_pmt_growth(0.05, 0.02, 5) #0.2117597
}
