% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_pvfb}
\alias{get_pvfb}
\title{Calculate the Present Value of Future Benefits (PVFB)}
\usage{
get_pvfb(sep_rate_vec, interest_vec, value_vec)
}
\arguments{
\item{sep_rate_vec}{Numeric vector. The annual separation rates for each future period.}

\item{interest_vec}{Numeric vector. The annual interest (discount) rates use in that period.}

\item{value_vec}{Numeric vector. The future benefits payable at each period.}
}
\value{
A numeric vector of the same length as `value_vec`, where each element represents the
        present value of future benefits starting from that period.
}
\description{
Given a vector of separation rates, a vector of corresponding interest rates, and a vector of
future values (benefits), this function computes the present value of these benefits. At each
period `i`, it calculates the probability of separation and then discounts the subsequent
future values back to the present using the provided interest rates.
}
\examples{
sep_rate_vec <- c(0.01, 0.02, 0.03, 0.04)
interest_vec <- c(0.05, 0.05, 0.05, 0.05)
value_vec <- c(100, 200, 300, 400)
get_pvfb(sep_rate_vec, interest_vec, value_vec)
}
