% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{npv}
\alias{npv}
\title{npv calculates the present value of future cashflows (cf)}
\usage{
npv(rate, cf)
}
\arguments{
\item{rate}{Annual discount rate (scalar double).}

\item{cf}{A list of cash flows in n=length(cf) number of years}
}
\value{
Numeric value.
}
\description{
npv calculates the present value of future cashflows (cf)
}
\examples{
cf <- c(100, 200, 300, 400, 500, 600)
npv(0.05, cf)  # 1704.37
}
