% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pv}
\alias{pv}
\title{Present value of a future payment}
\usage{
pv(rate, g = 0, nper, pmt, t = 1)
}
\arguments{
\item{rate}{Annual discount rate (scalar double).}

\item{g}{Growth rate (details needed).}

\item{nper}{Number of periods per year (scalar).}

\item{pmt}{Future payment, per period per year (scalar double).}

\item{t}{Number of years in the future (scalar).}
}
\value{
Numeric value.
}
\description{
Present value of a future payment
}
\examples{
pv(.1, g=0, nper=1, pmt=100, t=1) # 90.90909
pv(.1, g=0, nper=1, pmt=100, t=2) # 82.644628
pv(.1, g=0, nper=2, pmt=100, t=2) # 157.7761
pv(.1, g=.05, nper=2, pmt=100, t=2) # 161.5327
}
